Casting problem

+++ org.mwc.debrief.legacy/src/Debrief/GUI/Views/CoreCoordinateRecorder.java
@@ -349,7 +349,7 @@ public abstract class CoreCoordinateRecorder
           firstTime), new HiResDate(lastTime));
 
       // sort out a scale factor
-      final double scale = _worldIntervalMillis / _modelIntervalMillis;
+      final double scale = ((double)_worldIntervalMillis) / _modelIntervalMillis;
 
       final SimpleDateFormat df = new GMTDateFormat("ddHHmm.ss");
